feat(monkey): Move AddRemoveDcNemesis into own module
(cherry picked from commit 8dffb9bb3b87bf34a0c9d9106a97f6c66ce581d0)
